+++ packages/adapter-fastset/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @vleap/warps-adapter-fastset
 
+## 0.1.0-beta.56
+
+### Patch Changes
+
+- fa98f00: update wallets
+- Updated dependencies [fa98f00]
+  - @vleap/warps@3.0.0-beta.187
+
 ## 0.1.0-beta.55
 
 ### Patch Changes